sh: clean up some issues caught by shellcheck
The most important bits here are quoting the $@ invocations, especially in vg
globalExec; without this, trying to pass arguments containing spaces to the
program exec'd by vg globalExec will fail.  Some other instances are helpful if
your home directory has a space in its path. Some of the other quotes are
perhaps extraneous (eg, with echo), but it's a good habit to be in anyway.